refactor use installations
